Multiple fixes
