+++ backend/infrahub/core/validators/checks_runner.py
@@ -1,5 +1,5 @@
 import asyncio
-from typing import Coroutine
+from typing import Any, Coroutine
 
 from infrahub_sdk.node import InfrahubNode
 
@@ -7,7 +7,9 @@ from infrahub.core.constants import ValidatorConclusion, ValidatorState
 from infrahub.core.timestamp import Timestamp
 
 
-async def run_checks_and_update_validator(checks: list[Coroutine], validator: InfrahubNode) -> None:
+async def run_checks_and_update_validator(
+    checks: list[Coroutine[Any, None, ValidatorConclusion]], validator: InfrahubNode
+) -> None:
     """
     Execute a list of checks coroutines, and set validator fields accordingly.
     Tasks are retrieved by completion order so as soon as we detect a failing check,
@@ -22,7 +24,7 @@ async def run_checks_and_update_validator(checks: list[Coroutine], validator: In
 
     for earliest_task in asyncio.as_completed(checks):
         result = await earliest_task
-        if result == ValidatorConclusion.FAILURE:
+        if validator.conclusion.value != ValidatorConclusion.FAILURE.value and result == ValidatorConclusion.FAILURE:
             validator.conclusion.value = ValidatorConclusion.FAILURE.value
             await validator.save()
             # Continue to iterate to wait for the end of all checks
